Reorganize: Extract Statement and its impl into its own module.
